+++ app/services/create_order_service.rb
@@ -11,7 +11,7 @@ module CreateOrderService
         price_cents: artwork_price(artwork, edition_set_id: edition_set_id),
         quantity: quantity
       )
-      RejectExpiredOrdersJob.set(wait_until: order.state_expires_at).perform_later(order.id, order.state)
+      ExpireOrderJob.set(wait_until: order.state_expires_at).perform_later(order.id, order.state)
       order
     end
   end
