+++ app/services/order_service.rb
@@ -50,7 +50,7 @@ module OrderService
       TransactionService.create!(order, transaction)
       order.save!
       PostNotificationJob.perform_later(order.id, Order::APPROVED, by)
-      RejectExpiredOrdersJob.set(wait_until: order.state_expires_at).perform_later(order.id, order.state)
+      ExpireOrderJob.set(wait_until: order.state_expires_at).perform_later(order.id, order.state)
     end
     order
   rescue Errors::PaymentError => e
