+++ app/services/order_submit_service.rb
@@ -18,7 +18,7 @@ module OrderSubmitService
       order.transaction_fee_cents = calculate_transaction_fee(order)
       order.save!
       PostNotificationJob.perform_later(order.id, Order::SUBMITTED, by)
-      RejectExpiredOrdersJob.set(wait_until: order.state_expires_at).perform_later(order.id, order.state)
+      ExpireOrderJob.set(wait_until: order.state_expires_at).perform_later(order.id, order.state)
     end
 
     order
