+++ spec/services/create_order_service_spec.rb
@@ -36,7 +36,7 @@ describe CreateOrderService, type: :services do
             expect(order.line_items.first.artwork_id).to eq 'artwork-id'
             expect(order.line_items.first.edition_set_id).to eq 'edition-set-id'
             expect(order.line_items.first.quantity).to eq 2
-            expect(RejectExpiredOrdersJob).to have_been_enqueued.at(order.reload.state_expires_at)
+            expect(ExpireOrderJob).to have_been_enqueued.at(order.reload.state_expires_at)
           end.to change(Order, :count).by(1).and change(LineItem, :count).by(1)
         end
       end
